Fix build in Opensolaris 2009.06
Explicitly cast some values to their respective native types during printf
and workaround an incompatibility with false.  Fixes build and 'make check'
in Opensolaris 2009.06.

Monotone-Parent: 1bf57d13b7efd58f1c630e6c634cb1bd14ed65ba
Monotone-Revision: 1eb38615303d1486d1d1bceef6fdad2f33c9d2a3

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2009-11-23T09:29:57
Monotone-Branch: org.NetBSD.atf.src

+++ atf-c/fs.c
@@ -101,7 +101,7 @@ invalid_umask_format(const atf_error_t err, char *buf, size_t buflen)
     snprintf(buf, buflen, "Could not create the temporary %s %s because "
              "it will not have enough access rights due to the current "
              "umask %05o", stat_type_to_string(data->m_type),
-             data->m_path, data->m_umask);
+             data->m_path, (unsigned int)data->m_umask);
 }
 
 static
